♻️ refactor: share usage value formatter
