+++ packages/utils/src/format.test.ts
@@ -13,6 +13,7 @@ import {
   formatSpeed,
   formatTime,
   formatTokenNumber,
+  formatUsageValue,
 } from './format';
 
 describe('format', () => {
@@ -251,6 +252,31 @@ describe('format', () => {
       expect(formatTokenNumber(1048576)).toBe('1M'); // Gemini Flash
       expect(formatTokenNumber(2000000)).toBe('2M');
       expect(formatTokenNumber(2097152)).toBe('2M'); // Gemini Pro
+    });
+  });
+
+  describe('formatUsageValue', () => {
+    it('formats token usage details with short units', () => {
+      expect(formatUsageValue(93_405)).toBe('93.4K');
+      expect(formatUsageValue(92_119)).toBe('92.1K');
+      expect(formatUsageValue(3_488)).toBe('3.5K');
+      expect(formatUsageValue(189_018)).toBe('189K');
+    });
+
+    it('formats credit usage details with the same short units', () => {
+      expect(formatUsageValue(16_127)).toBe('16.1K');
+      expect(formatUsageValue(16_179)).toBe('16.2K');
+    });
+
+    it('keeps small token counts readable without suffixes', () => {
+      expect(formatUsageValue(0)).toBe('0');
+      expect(formatUsageValue(6)).toBe('6');
+      expect(formatUsageValue(999)).toBe('999');
+    });
+
+    it('formats million-level token counts with M suffix', () => {
+      expect(formatUsageValue(1_000_000)).toBe('1M');
+      expect(formatUsageValue(1_500_000)).toBe('1.5M');
     });
   });
