+++ packages/utils/src/format.ts
@@ -77,6 +77,12 @@ export const formatIntergerNumber = (num?: any) => {
   return numeral(num).format('0,0');
 };
 
+export const formatUsageValue = (number: number) => {
+  if (number >= 1_000_000) return `${numeral(number / 1_000_000).format('0.[0]')}M`;
+  if (number >= 1_000) return `${numeral(number / 1_000).format('0.[0]')}K`;
+  return numeral(number).format('0,0');
+};
+
 export const formatTokenNumber = (num: number): string => {
   if (!num && num !== 0) return '--';
