+++ src/locales/default/chat.ts
@@ -188,7 +188,7 @@ export default {
   'input.costEstimate.hint': 'Estimated cost: ~${{cost}}',
   'input.costEstimate.settingsLink': 'Adjust warning threshold',
   'input.costEstimate.tooltip':
-    'Estimated from the current context, enabled tools, and model pricing. Input: {{inputTokens}} tokens. Estimated output: {{outputTokens}} tokens. Actual cost may vary.',
+    'Estimated from current context, tools, and model pricing. Input ~{{inputTokens}}, output ~{{outputTokens}}. Actual cost may vary.',
   'input.disclaimer': 'Agents can make mistakes. Use your judgment for critical info.',
   'input.errorMsg': 'Send failed: {{errorMsg}}. Retry, or send again later.',
   'input.more': 'More',
